Add files via upload
The objective of this exercise was to build a model using a Supervised learning technique to figure out profitable segments to target for cross-selling personal loans. A Pilot campaign data of 20000 customers was used which included several demographic and behavioral variables. The Model was further validated and a deployment strategy was recommended.